refactor: Reorganized lambdas folder structure by services
