+++ src/lambdas/index.ts
@@ -1,3 +1,3 @@
-export { handler as randomGet } from './random-get'
+export { handler as randomGet } from './services/dogBreed-api/randomBreed/random-get'
 
-export { handler as breedsGet } from './breeds-get'
+export { handler as breedsGet } from './services/dogBreed-api/breedsAll/breeds-get'
